Move ADFS-related handler stuff to own package
